# generated/openapi/notes-app-api.openapi.json
{
  "openapi": "3.1.2",
...
  "info": {
    "title": "SDKWork Notes App API",
    "version": "0.1.0",
    "description": "App/client authority skeleton for SDKWork Notes page workspaces. Durable page bytes, folders, assets, and versions are Drive-backed; Notes owns page metadata, object model, search, sync, and AI governance metadata."
  },
  "servers": [
    {
      "url": "/app/v3/api",
      "description": "SDKWork App API root"
    }
  ],
...
  "paths": {
...
    "/app/v3/api/notes/pages/{pageId}/versions/{driveVersionId}/restore": {
      "post": {
        "summary": "Restore a Drive-owned page content version through the Notes business facade.",
        "operationId": "pages.versions.restore",
        "tags": [
          "notes"
        ],
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageIdPath"
          },
          {
            "$ref": "#/components/parameters/DriveVersionIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePageVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restored page content envelope with the new current Drive version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageContent"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        },
        "x-sdkwork-owner": "sdkwork-notes",
        "x-sdkwork-api-authority": "sdkwork-notes-app-api",
        "x-sdkwork-domain": "notes",
        "x-sdkwork-resource": "pages",
        "x-sdkwork-permission": "notes.pages.versions.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notes.page.versions.restore",
        "x-sdkwork-idempotent": false
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "parameters": {
...
      "PageIdPath": {
        "name": "pageId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Public Notes page id."
      },
...
      "DriveVersionIdPath": {
        "name": "driveVersionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Drive-owned node version id exposed through the Notes page facade."
      }
    },
    "responses": {
      "ProblemDetailResponse": {
        "description": "Problem detail error.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      }
    },
    "schemas": {
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "PageContent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "pageId",
          "driveNodeId",
          "driveVersionId",
          "driveVersionNo",
          "contentType",
          "content",
          "contentSchemaVersion"
        ],
        "properties": {
          "pageId": {
            "type": "string"
          },
          "driveNodeId": {
            "type": "string"
          },
          "driveVersionId": {
            "type": "string"
          },
          "driveVersionNo": {
            "type": "string"
          },
          "contentType": {
            "type": "string",
            "maxLength": 255
          },
          "contentSchemaVersion": {
            "type": "string",
            "maxLength": 32
          },
          "content": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "RestorePageVersionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantId",
          "organizationId",
          "operatorId"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "operatorId": {
            "type": "string"
          },
          "expectedCurrentDriveVersionId": {
            "type": "string"
          }
        }
      }
    }
  }
}